pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "rice_cnn_classifier"
version = "0.0.1"
description = "Classification of rice varieties using CNN."
authors = [
  { name = "Group 1", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc-gs>=3.0.2",
    "dvc[gs]>=3.0.0",
    "fastapi==0.115.6",
    "google-cloud-aiplatform>=1.73.0",
    "invoke>=2.2.0",
    "pathspec>=0.11.0,<0.12.0",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "pillow>=10.0.0",
    "torchvision>=0.17.0",
    "numpy>=1.26.0",
    "wandb>=0.24.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pydantic>=2.12.5",
    "gradio>=6.3.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "pytest-cov>=6.3.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*", "*/_remote_module_non_scriptable.py"]
ignore_errors = true
